requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
accelerate==1.5.2
    # via stableguard (pyproject.toml)
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.11.14
    # via
    #   datasets
    #   fsspec
aiosignal==1.3.2
    # via aiohttp
attrs==25.3.0
    # via aiohttp
certifi==2025.1.31
    # via requests
charset-normalizer==3.4.1
    # via requests
colorama==0.4.6
    # via tqdm
datasets==3.5.0
    # via stableguard (pyproject.toml)
diffusers==0.32.2
    # via stableguard (pyproject.toml)
dill==0.3.8
    # via
    #   datasets
    #   multiprocess
filelock==3.18.0
    # via
    #   datasets
    #   diffusers
    #   huggingface-hub
    #   torch
    #   transformers
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2024.12.0
    # via
    #   datasets
    #   huggingface-hub
    #   torch
huggingface-hub==0.29.3
    # via
    #   accelerate
    #   datasets
    #   diffusers
    #   tokenizers
    #   transformers
idna==3.10
    # via
    #   requests
    #   yarl
importlib-metadata==8.6.1
    # via diffusers
jinja2==3.1.6
    # via torch
markupsafe==3.0.2
    # via jinja2
mpmath==1.3.0
    # via sympy
multidict==6.2.0
    # via
    #   aiohttp
    #   yarl
multiprocess==0.70.16
    # via datasets
networkx==3.4.2
    # via torch
numpy==2.2.4
    # via
    #   accelerate
    #   datasets
    #   diffusers
    #   opencv-python
    #   pandas
    #   torchvision
    #   transformers
opencv-python==4.11.0.86
    # via stableguard (pyproject.toml)
packaging==24.2
    # via
    #   accelerate
    #   datasets
    #   huggingface-hub
    #   transformers
pandas==2.2.3
    # via datasets
pillow==11.1.0
    # via
    #   stableguard (pyproject.toml)
    #   diffusers
    #   pillow-heif
    #   torchvision
pillow-avif-plugin==1.5.1
    # via stableguard (pyproject.toml)
pillow-heif==0.22.0
    # via stableguard (pyproject.toml)
propcache==0.3.1
    # via
    #   aiohttp
    #   yarl
psutil==7.0.0
    # via accelerate
pyarrow==19.0.1
    # via datasets
python-dateutil==2.9.0.post0
    # via pandas
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via
    #   accelerate
    #   datasets
    #   huggingface-hub
    #   transformers
regex==2024.11.6
    # via
    #   diffusers
    #   transformers
requests==2.32.3
    # via
    #   datasets
    #   diffusers
    #   huggingface-hub
    #   transformers
safetensors==0.5.3
    # via
    #   accelerate
    #   diffusers
    #   transformers
setuptools==78.1.0
    # via torch
six==1.17.0
    # via python-dateutil
sympy==1.13.1
    # via torch
tokenizers==0.21.1
    # via transformers
torch==2.6.0+cu126
    # via
    #   stableguard (pyproject.toml)
    #   accelerate
    #   torchaudio
    #   torchvision
torchaudio==2.6.0+cu126
    # via stableguard (pyproject.toml)
torchvision==0.21.0+cu126
    # via stableguard (pyproject.toml)
tqdm==4.67.1
    # via
    #   datasets
    #   huggingface-hub
    #   transformers
transformers==4.50.2
    # via stableguard (pyproject.toml)
typing-extensions==4.13.0
    # via
    #   huggingface-hub
    #   torch
tzdata==2025.2
    # via pandas
urllib3==2.3.0
    # via requests
xxhash==3.5.0
    # via datasets
yarl==1.18.3
    # via aiohttp
zipp==3.21.0
    # via importlib-metadata
